+++ Android/app/src/main/AndroidManifest.xml
@@ -5,11 +5,6 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
-        android:maxSdkVersion="32"
-        tools:ignore="ScopedStorage" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"
-        android:maxSdkVersion="32" />
 
     <queries>
         <package android:name="com.topjohnwu.magisk" />
@@ -24,14 +19,15 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:requestLegacyExternalStorage="true"
-        android:theme="@style/AppTheme"
-        android:windowSoftInputMode="adjustResize"
+        android:theme="@style/AppTheme.SplashScreen"
         tools:ignore="GoogleAppIndexingWarning"
         tools:targetApi="q">
         <activity
             android:name="com.dergoogler.mmrl.MainActivity"
             android:configChanges="orientation|screenSize"
-            android:exported="true">
+            android:windowSoftInputMode="adjustResize"
+            android:exported="true"
+            tools:ignore="Instantiatable">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
