+++ Android/app/src/main/res/values/styles.xml
@@ -1,8 +1,8 @@
 <resources>
-    <style name="Theme.App.SplashScreen" parent="Theme.SplashScreen.IconBackground">
+    <style name="AppTheme.SplashScreen" parent="Theme.SplashScreen.IconBackground">
         <item name="windowSplashScreenAnimatedIcon">@drawable/ic_cdv_splashscreen</item>
         <item name="windowSplashScreenAnimationDuration">200</item>
-        <item name="postSplashScreenTheme">@style/Theme.AppCompat.NoActionBar</item>
+        <item name="postSplashScreenTheme">@style/AppTheme</item>
     </style>
 
     <style name="AppTheme" parent="Theme.AppCompat.DayNight.NoActionBar">
